Simplify the import path(s)
We should decouple the git repository from the import path

This will enable us to use a nice looking:
"openstack.org/golang-client/identity/v2"
instead of the:
"git.openstack.org/openstack/golang-client.git/identity/v2"

We drop the hostname, we drop the .git etc. Also, we can move
the git repository around and not affect the imports of
consuming projects.

Change-Id: I87d1d9ed73100401e4669f2bdb19ef7243fb00b4